Execution driven by an iterator that returns closures.
Using a combination of a) the raw list of toml keys, in order, top to
bottom, and b) the deserialized Config object, we can build a sequential
execution plan of []TypeIndex.

We wrap this plan in an iterator object, which also holds the
deserialized config. The plan is simply a list of indexes to different
parts of our config struct, sequenced in the order they appear in the
file.

This is dead simple file ordering. We will focus on making this solid
and stable, and we can enhance the execution ordering with dependent
states later.

Refs #1